requirements.txt
absl-py==2.1.0
aiohappyeyeballs
aiohttp
aiosignal==1.3.1
anyio==4.6.2.post1
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
bitsandbytes
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
chromadb==1.5.0
dill==0.3.8
Django==5.1.15
django-environ==0.11.2
djangorestframework==3.15.2
docstring_parser==0.16
filelock==3.20.3
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.0
h11==0.16.0
h5py==3.12.1
hf_transfer==0.1.8
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
networkx==3.4.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow
propcache==0.2.0
protobuf==5.29.6
psutil
pyarrow==18.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
regex==2024.9.11
safetensors==0.4.5
sentencepiece==0.2.1
setuptools==78.1.1
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
sympy==1.13.1
termcolor==2.5.0
tqdm==4.66.6
torch
torchvision
torchaudio
transformers==5.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.6.3
Werkzeug==3.1.5
wheel==0.46.2
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.2
Crawl4AI==0.8.0
unstructured
langchain-openai==1.1.7
Markdown
spacy
onnxruntime
